Initial Source Files
Just laying out the ground-work a bit for what will eventually be actual
source code that does stuff and what not. Most of the files are empty'ish
right now. The config class does represent the only config that exists as
of right now ('port') and the actual config (reference.json) will act as
the base/main config (not sure how to handle overrides just yet).