Added init script and route-matcher config
The route matcher config will live external to the jars (thus the
project). So, I've created an etc folder (not packged in the jar)
that stores a sample config (for development purposes). I've also
added an init script which will install (sym-link) the default file
to the default-location on the machine.