Re-Defined config auto-loading
Makes sense to only re-load the route-configurations and not the base-
config. For the base config, we only need to provide overrides. The
config-loader is now setup to load the base config with overrides (if
specified) and then schedule the routes to be loaded.

This means that when we boot we only ensure that we can load the app-
config and the route-configs will be loaded later.